16 bytes shaved off PRG: 8140 → 8124 - diff:  16
Total size reduction: 16 bytes

+++ SAUCE/functions/collision.h
@@ -864,7 +864,7 @@ void tmp20f() {
 }
 
 void commonly_used_death_check() {
-	if ((temp_x & 0x0f) >= 0x04 && (temp_x & 0x0f) < 0x09) {
+	if ((uint8_t)(temp_x & 0x0f) >= 0x04 && (uint8_t)(temp_x & 0x0f) < 0x09) {
 		cube_data[currplayer] = 1;
 	}
 }
